+++ packages/base-modal/src/matches-polyfill.ts
@@ -17,3 +17,5 @@ if (typeof window !== 'undefined') {
             };
     }
 }
+
+export {};