+++ packages/select/src/components/base-select/components/list-desktop/types/types.ts
@@ -1,3 +1,0 @@
-import { getListPopoverDesktopProps } from '../helpers/get-list-popover-desktop-props';
-
-export type ListPopoverDesktopRestProps = ReturnType<typeof getListPopoverDesktopProps>;