+++ packages/tabs/src/components/secondary-tablist/Component.desktop.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import cn from 'classnames';
 
 import { TagDesktop } from '@alfalab/core-components-tag/desktop';
 
@@ -6,20 +7,26 @@ import { SecondaryTabListProps } from '../../typings';
 
 import { SecondaryTabList } from './Component';
 
-import commonStyles from './index.module.css';
+import styles from './index.module.css';
 
 export type SecondaryTabListDesktopProps = Omit<SecondaryTabListProps, 'tagSize' | 'breakpoint'>;
 
 export const SecondaryTabListDesktop = ({
+    className,
     size = 's',
+    tagView,
     ...restProps
 }: SecondaryTabListDesktopProps) => (
     <SecondaryTabList
         {...restProps}
         TagComponent={TagDesktop}
         size={size}
-        styles={commonStyles}
+        styles={styles}
+        className={cn(className, styles.desktop, {
+            [styles.transparentView]: tagView === 'transparent',
+        })}
         tagSize={size}
+        tagView={tagView}
         platform='desktop'
     />
 );