+++ packages/tooltip/src/component.screenshots.test.tsx
@@ -6,7 +6,7 @@ import {
     matchHtml,
     closeBrowser,
     createPreview,
-} from '../../screenshot-utils';
+} from '@alfalab/core-components-screenshot-utils';
 
 describe('Tooltip', () => {
     createPreview(