+++ packages/universal-modal/src/desktop/types/props.ts
@@ -43,11 +43,6 @@ export type BaseUniversalModalProps = {
      * Устанавливает отступы модального окна
      */
     margin?: TMargin;
-
-    /**
-     * Хэндлер закрытия модалки
-     */
-    onClose?: () => void;
 };
 
 export type UniversalModalDesktopProps = BaseUniversalModalProps &
@@ -62,6 +57,8 @@ export type UniversalModalDesktopProps = BaseUniversalModalProps &
             | 'onUnmount'
             | 'transitionProps'
             | 'backdropProps'
+            | 'disableBackdropClick'
+            | 'onClose'
         >
     > & {
         /**