+++ tools/gh-actions/.eslintrc.cjs
@@ -1,0 +1,10 @@
+/**
+ * @type {import('eslint').Linter.Config}
+ */
+const config = {
+    rules: {
+        '@typescript-eslint/dot-notation': 'off',
+    },
+};
+
+module.exports = config;