+++ packages/textarea/CHANGELOG.md
@@ -1,4 +1,41 @@
-# Change Log
+# @alfalab/core-components-textarea
+
+## 9.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-form-control@13.0.0
+    -   @alfalab/core-components-mq@5.0.0
+    -   @alfalab/core-components-shared@1.0.0
+
+## 8.9.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.18.0
+    -   form-control@12.8.6
+
+## 8.9.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.1
+    -   form-control@12.8.5
 
 ## 8.9.4